gcc: fix warnings in sys_*_iram

+++ library/gcc/include/sys/bank.h
@@ -51,7 +51,7 @@ static inline uint16_t bank_get_map(uint16_t bank_type) {
 		: "b" (bank_type), "Rah" ((uint8_t) 0x01)
 		: "cc", "memory"
 	);
-	return result;
+	return (uint16_t) result;
 }
 
 /**
@@ -67,7 +67,7 @@ static inline uint8_t bank_read_byte(uint16_t bank_id, uint16_t offset) {
 		: "b" (bank_id), "d" (offset), "Rah" ((uint8_t) 0x02)
 		: "cc", "memory"
 	);
-	return result;
+	return (uint8_t) result;
 }
 
 /**
@@ -98,7 +98,7 @@ static inline uint8_t bank_read_word(uint16_t bank_id, uint16_t offset) {
 		: "b" (bank_id), "d" (offset), "Rah" ((uint8_t) 0x04)
 		: "cc", "memory"
 	);
-	return result;
+	return (uint8_t) result;
 }
 
 /**
